+++ Sources/CurrentDevice.swift
@@ -24,8 +24,10 @@ public extension String {
 public extension Device {
     /// An object representing the current device this software is running on.
     // TODO: Figure out how to handle this with concurrency.
+    @available(iOS 13.0, tvOS 13, watchOS 6, *)
     @MainActor
     static let current: some CurrentDevice = ActualHardwareDevice() // singleton representing the current device but separated so that we can replace or mock
+    @available(iOS 13.0, tvOS 13, watchOS 6, *)
     enum Environment: DeviceAttributeExpressible, Sendable { // unable to conform to CaseIterable since @MainActor isolated
         case realDevice, simulator, playground, preview, designedForiPad, macCatalyst
         
@@ -147,6 +149,7 @@ public extension ProcessInfo.ThermalState {
 //#if canImport(Observable)
 //@Observable
 //#endif // TODO: this is only supported in iOS 17+ so wait to implement until we no longer need backwards compatibility
+@available(iOS 13.0, tvOS 13, watchOS 6, *)
 @MainActor // Inferred conformance to Sendable
 public protocol CurrentDevice: ObservableObject, DeviceType, Sendable { // needs explicit conformance to for background await usage.
     associatedtype BatteryType: Battery
@@ -220,6 +223,7 @@ public protocol CurrentDevice: ObservableObject, DeviceType, Sendable { // needs
     /// will enable monitoring at the specified frequency.  If this is called multiple times, it will replace the existing monitor.
     func enableMonitoring(frequency: TimeInterval)
 }
+@available(iOS 13.0, tvOS 13, watchOS 6, *)
 extension CurrentDevice {
     /// Returns the version number of Swift being used to compile
     public var swiftVersion: String {
@@ -290,6 +294,7 @@ Device Framework Version: v\(Device.version)
 }
 
 // this is internal because it shouldn't be directly needed outside the framework.  Everything is exposed via CurrentDevice protocol.
+@available(iOS 13.0, tvOS 13, watchOS 6, *)
 @MainActor // All calculations/queries should be quick so we can isolate to main actor to give Sendable conformance.
 final class ActualHardwareDevice: CurrentDevice {
 #if canImport(Combine)
@@ -387,7 +392,7 @@ final class ActualHardwareDevice: CurrentDevice {
         }
 #if canImport(Combine)
         // Note: this will be "false" under Catalyst which is what we want.
-        if #available(watchOS 7.0, *) { // not available on watchOS 6
+        if #available(iOS 14, watchOS 7.0, macOS 11, tvOS 14, *) { // not available on watchOS 6
             return ProcessInfo.processInfo.isiOSAppOnMac
         }
 #endif
@@ -498,7 +503,7 @@ final class ActualHardwareDevice: CurrentDevice {
 //        print("UIDevice.current.systemName: \(systemName)")
         let systemVersion = Version(UIDevice.current.systemVersion)
 //        print("UIDevice.current.systemVersion: \(systemVersion)")
-        if idiom == .pad, #available(iOS 13, *), systemName == "iOS" {
+        if idiom == .pad, systemName == "iOS" {
             systemName = "iPadOS"
 //            print("System Version changed to: \(systemName)")
         }
@@ -826,6 +831,7 @@ final class ActualHardwareDevice: CurrentDevice {
     
 }
 
+@available(iOS 13.0, tvOS 13, watchOS 6, *)
 @MainActor
 public final class MockDevice: CurrentDevice {
     public typealias BatteryType = MockBattery
@@ -1076,7 +1082,7 @@ public final class MockDevice: CurrentDevice {
 
 #if canImport(SwiftUI) && swift(>=5.9)
 import SwiftUI
-@available(watchOS 8.0, tvOS 15.0, macOS 12.0, *)
+@available(iOS 15, watchOS 8.0, tvOS 15.0, macOS 12.0, *)
 #Preview("Animated Test") {
     List {
         CurrentDeviceInfoView(device: Device.current, includeStorage: true)