+++ Sources/DeviceInfoView.swift
@@ -13,110 +13,8 @@ public extension CGFloat {
     static let devicePanelRadius: Double = 15
 }
 
-// Normally would just use KuditFrameworks but just in case that isn't available...
-extension Color {
-    init(hex: String) {
-        let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-        var int: UInt64 = 0
-        Scanner(string: hex).scanHexInt64(&int)
-        let a, r, g, b: UInt64
-        switch hex.count {
-        case 3: // RGB (12-bit)
-            (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
-        case 6: // RGB (24-bit)
-            (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
-        case 8: // ARGB (32-bit)
-            (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
-        default:
-            (a, r, g, b) = (1, 1, 1, 0)
-        }
-        
-        self.init(
-            .sRGB,
-            red: Double(r) / 255,
-            green: Double(g) / 255,
-            blue:  Double(b) / 255,
-            opacity: Double(a) / 255
-        )
-    }
-}
-
-// for switching between asset images and systemImages
-public extension Image {
-    /// Create image with a symbol name using system SF symbol or fall back to the symbol asset embedded in Device library.
-    init(symbolName: String) {
-        var symbolName = symbolName
-        let legacySymbolName = "\(symbolName).legacy"
-        // use the new symbol name for the Xcode 15 symbol assets (should include colors and proper layering)
-        if #available(iOS 17.0, watchOS 10.0, macOS 14.0, tvOS 17.0, visionOS 1.0, macCatalyst 17.0, *) {
-            symbolName = symbolName.safeSymbolName(fallback: legacySymbolName)
-        } else {
-            // if older OS, fallback to compatible symbols.
-            symbolName = legacySymbolName.safeSymbolName(fallback: symbolName)
-        }
-        if .nativeSymbolCheck(symbolName) {
-            self.init(systemName: symbolName)
-        } else {
-            // get module image asset if possible
-            self.init(symbolName, bundle: Bundle.module)
-        }
-    }
-    @MainActor
-    init(_ symbolRepresentable: some SymbolRepresentable) {
-        self.init(symbolName: symbolRepresentable.symbolName)
-    }
-}
-// TODO: Move the following to Kudit Frameworks
-extension String {
-    public static let defaultFallback = "questionmark.square.fill"
-    /*
-     Legacy versions for Symbol (iOS = catalyst = tvOS
-     Device min: 15, 11, 14, 6 so create 1.0 or 2.0 versions for fallback.  Make note that watchOS 6 doesn’t support new symbols.
-     1.0 = iOS 13, macOS 11, watchOS 6 * Check this with Device minimum version for potential fallbacks or put note that symbols only work on iOS 13+
-     2.0 = iOS 14, macOS 11, watchOS 7, Xcode 12
-     3.0 = iOS 15, macOS 12, watchOS 8, Xcode 13
-     4.0 = iOS 16, macOS 13, watchOS 9, Xcode 14
-     5.0 = iOS 17, macOS 14, watchOS 10, Xcode 15 * Anything before this, use legacy version.
-     */
-    /// helper for making sure symbolName: function always returns an actual image and never `nil`.
-    public func safeSymbolName(fallback: String = .defaultFallback) -> String {
-        if !.nativeSymbolCheck(self) {
-            // check for asset
-            if !.nativeLocalCheck(self) {
-                if fallback == .defaultFallback {
-                    return fallback
-                } else {
-                    // go through the fallback symbol to make sure it's valid (should never really happen)
-                    return fallback.safeSymbolName()
-                }
-            }
-        }
-        return self
-    }
-}
-extension Bool {
-    static func nativeSymbolCheck(_ symbolName: String) -> Bool {
-#if canImport(UIKit)
-        return UIImage(systemName: symbolName) != nil
-#elseif canImport(AppKit)
-        return NSImage(systemSymbolName: symbolName, accessibilityDescription: nil) != nil
-#endif
-    }
-    static func nativeLocalCheck(_ symbolName: String) -> Bool {
-#if canImport(UIKit)
-        return UIImage(named: symbolName, in: Bundle.module, with: nil) != nil
-#elseif canImport(AppKit)
-        if #available(macOS 13, *) {
-            return NSImage(symbolName: symbolName, bundle: Bundle.module, variableValue: 1) != nil
-        } else {
-            // probably won't work in macOS 12
-            return NSImage(named: symbolName) != nil
-        }
-#endif
-    }
-}
-
 /// Technincally a function but present like a View struct.
+@available(iOS 13, tvOS 13, watchOS 6, *)
 @MainActor
 public func CapabilitiesTextView(capabilities: Capabilities) -> Text {
     var output = Text("")
@@ -135,13 +33,19 @@ public func CapabilitiesTextView(capabilities: Capabilities) -> Text {
         else if case .watchSize = capability {}
         else {
             if #available(watchOS 7.0, *) {
-                if #available(macOS 12.0, watchOS 8.0, tvOS 15.0, *) {
-                    output = output + Text(Image(symbolName: capability.symbolName).symbolRenderingMode(.hierarchical)
-                    ) + Text(" ")
+                // Fallback on earlier versions
+                if #available(iOS 14, macOS 11.0, tvOS 14, *) {
+                    if #available(iOS 15, macOS 12.0, watchOS 8.0, tvOS 15.0, *) {
+                        output = output + Text(Image(symbolName: capability.symbolName).symbolRenderingMode(.hierarchical)
+                        ) + Text(" ")
+                    } else {
+                        output = output + Text(Image(capability)
+                        ) + Text(" ")
+                    }
                 } else {
                     // Fallback on earlier versions
-                    output = output + Text(Image(capability)
-                    ) + Text(" ")
+                    output = output + Text(capability.caseName)
+                    + Text(" ")
                 }
             } else {
                 // Fallback on earlier versions
@@ -152,7 +56,7 @@ public func CapabilitiesTextView(capabilities: Capabilities) -> Text {
     return output
 }
 
-@available(watchOS 8.0, tvOS 15.0, macOS 12.0, *)
+@available(iOS 15, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 @MainActor
 public struct DeviceInfoView: View {
     public var device: DeviceType
@@ -181,8 +85,13 @@ public struct DeviceInfoView: View {
                     HStack {
                         deviceColors
                         Spacer(minLength: 0)
-                        Text("\(device.cpu.caseName) ").font(.footnote.smallCaps())+Text(
-                            Image(symbolName: "cpu"))
+                        if #available(iOS 14.0, *) {
+                            Text("\(device.cpu.caseName) ").font(.footnote.smallCaps())+Text(
+                                Image(symbolName: "cpu"))
+                        } else {
+                            // Fallback on earlier versions
+                            Text("\(device.cpu.caseName) ").font(.footnote.smallCaps())+Text("cpu")
+                        }
                     }
                     HStack {
                         CapabilitiesTextView(capabilities: device.capabilities)
@@ -213,7 +122,7 @@ public struct DeviceInfoView: View {
     }
 }
 
-@available(watchOS 8.0, tvOS 15.0, macOS 12.0, *)
+@available(iOS 15, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 public struct DeviceListView: View {
     public var devices: [DeviceType]
     
@@ -269,7 +178,7 @@ public struct DeviceListView: View {
     }
 }
 
-@available(watchOS 8.0, tvOS 15.0, macOS 12.0, *)
+@available(iOS 15, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 #Preview("All Devices") {
     NavigationView {
         DeviceListView(devices: Device.all)